Add utilities for bootstrap predictions and enhance visualization functions
- Implemented `generate_bootstrap_predictions` in `utilities.py` to generate bootstrap predictions with confidence intervals.
- Created a comprehensive set of visualization functions in `visualization.py` for analyzing HIV case trends, model fits, ensemble comparisons, and performance metrics.
- Added decorators for managing plot saving and closing.
- Included functionality for forecasting future trends with uncertainty quantification and saving results to CSV.
